+++ server/src/models/index.js
@@ -1,7 +1,6 @@
 import fs from 'fs';
 import path from 'path';
 import Sequelize from 'sequelize';
-// import { config as configuration }  from '../config/config.json';
 
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
